tls_core: avoid infinite loops in cleanup
In two places in `tls_core.c` if there is an error control flow is
redirected to the `cleanup` label to log the error/perform cleanup
tasks.

However, the body of the cleanup label logic in both cases includes its
own `goto cleanup` at the end. Since the `rustls_result` guarding the
logic won't have changed since the original error it seems like this
will produce an infinite loop, forever jumping back to the start of the
`cleanup` section.

This commit removes the `goto cleanup` from the `cleanup` labelled
blocks.