ci: install w/ cargo-c instead of cmake
cmake is only required for the client/server example binaries. For
mod_tls we only need the librustls bits and can use cargo-c directly.

cargo-c respects the host OS pattern of putting libs in an
arch-specific subdir. To make the older 0.14.x Makefile install and the
cargo-c install equivalent we override this behaviour by setting
--libdir.