ci: restore aws-lc-rs crypto provider in matrix
This works without issue with the main branch of rustls-ffi.

For the 0.14.1 release when using the Makefile to install there's an
upstream issue where the CFLAGS intended for the rustls-ffi
client/server examples are _also_ used when building librustls and its
rust dependencies.

This includes aws-lc-sys, which at the version used by
librustls/rustls/aws-lc-rs (0.21.1) produces errors under the pedantic
settings used by rustls-ffi for its examples. The warnings have been
fixed upstream, but not for the version in use by rustls-ffi 0.14.x.

To work around this issue we pass CFLAGS="" when running the make
install for librustls 0.14.1.